+++ app/javascript/components/widget/analysis-result/templates.js
@@ -2,6 +2,7 @@ import NumberOfServices from '../widget-templates/number-of-services';
 import MinimunDistance from '../widget-templates/minimun-distance';
 import AccessPointsPercentage from '../widget-templates/points-percentage';
 import AreaCoverage from '../widget-templates/area-coverage';
+import SexPercentage from '../widget-templates/sex-percentage';
 
 import SummaryWidget from '../summary-widget';
 
@@ -12,7 +13,7 @@ const widgetTemplates = new Map([
   ['Minimum distance to access points', { component: MinimunDistance }],
   ['Area covered by GSM, 3G, and 4G', { component: AreaCoverage }],
   ['Urban and Rural areas', { component: NumberOfServices }],
-  ['Population sex percentage', { component: AreaCoverage }],
+  ['Population sex percentage', { component: SexPercentage }]
 ]);
 
 export default widgetTemplates;