+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Added filters on the insights page
 - Added year selector inside legend sector layers.
 - Link to download widgetdata in CSV
+- Sex percentage widget for specific countries
 
 ### Changed